Add gitlab e2e test
