+++ test/testdata/pipeline_remote_annotations.yaml
@@ -4,18 +4,12 @@ kind: Pipeline
 metadata:
   name: pipeline-remote-annotations
 spec:
-  # workspaces:
-  # - name: source
-  # - name: basic-auth
-  # params:
-  #   - name: repo_url
-  #   - name: revision
   tasks:
     - name: task-spec
       taskSpec:
         steps:
           - name: task-spec
-            image: registry.access.redhat.com/ubi8/ubi-minimal:8.2
+            image: registry.access.redhat.com/ubi8/ubi-micro:8.5
             script: |
               echo "Hello from taskSpec"
               exit 0
